remove old coordinator
